# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kapu$ta app endpoints document",
    "version": "1.0.0",
    "title": "Swagger Kapu$ta",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "oksana_stegko@ukr.net"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "kapusta-final-project-nodejs.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Get all categories",
        "description": "This can only be done by the logged in user.",
        "operationId": "getAllCategories",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "category"
        ],
        "summary": "Add new category",
        "description": "This can only be done by the logged in user.",
        "operationId": "addCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "подарки"
                },
                "sign": {
                  "type": "string",
                  "example": "incomes"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "201 OK. Category added!"
                },
                "results": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "df5v1D"
                    },
                    "name": {
                      "type": "string",
                      "example": "подарки"
                    },
                    "sign": {
                      "type": "string",
                      "example": "incomes"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "category"
        ],
        "summary": "Update category",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateCategory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "new подарки"
                },
                "sign": {
                  "type": "string",
                  "example": "incomes"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "201 OK. Category updated!"
                },
                "results": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "df5v1D"
                    },
                    "name": {
                      "type": "string",
                      "example": "new подарки"
                    },
                    "sign": {
                      "type": "string",
                      "example": "incomes"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }

    },
    "/categories/{signValue}": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Get categories by sign",
        "description": "This can only be done by the logged in user.",
        "operationId": "getCategoriesBySign",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "signValue",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "200 OK."
                },
                "results": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Category"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
...
    "/reports/detals":{
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Get report details by transactions",
        "description": "This can only be done by the logged in user.",
        "operationId": "detalsTransactions",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name":"sign",
            "in":"query",
            "required":true,
            "schema":{
              "type":"string",
              "example":"incomes"
            }
          },
          {
            "name":"year",
            "in":"query",
            "required":true,
            "schema":{
              "type":"number",
              "example":2021
            }
          },
          {
            "name":"month",
            "in":"query",
            "required":true,
            "description":"Count of month start from 0. January - 0",
            "schema":{
              "type":"number",
              "example":10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report of the Details by transaction",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "200 OK. Report of the details!"
                },
                "category": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Category"
                  }
                },
                "date": {
                  "$ref":"#/definitions/Transaction/properties/date"
                },
                "description":{
                  "$ref":"#/definitions/Transaction/properties/description"
                },
                "value":{
                  "$ref":"#/definitions/Transaction/properties/value"
                },
                "month":{
                  "$ref":"#/definitions/Transaction/properties/month"
                },
                "year":{
                  "$ref":"#/definitions/Transaction/properties/year"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
    "/reports/group-by-category":{
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Get report details by transactions and grouping by categories",
        "description": "This can only be done by the logged in user.",
        "operationId": "groupByCategory",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name":"sign",
            "in":"query",
            "required":true,
            "schema":{
              "type":"string",
              "example":"incomes"
            }
          },
          {
            "name":"year",
            "in":"query",
            "required":true,
            "schema":{
              "type":"number",
              "example":2021
            }
          },
          {
            "name":"month",
            "in":"query",
            "description":"Count of month start from 0. January - 0",
            "required":true,
            "schema":{
              "type":"number",
              "example":10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report of the Details by transaction and then grouping by categories",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "200 OK. Report of the details with grouping them later!"
                },
                "results": {
                  "type": "array",
                  "items": {
                  "type":"object",
                  "properties":{
                    "categoryId":{
                      "$ref":"#/definitions/Category/properties/_id"
                    },
                    "categoryName":{
                      "$ref":"#/definitions/Category/properties/name"
                    },
                    "value":{
                      "type":"number",
                      "example": 0
                    },
                    "subCategories":{
                      "type":"array",
                      "items":{
                        "type":"object",
                        "properties":{
                          "name":{
                            "$ref":"#/definitions/Transaction/properties/description"
                          },
                          "value":{
                            "$ref":"#/definitions/Transaction/properties/value"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        }},
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    },
    "/reports/year-report":{
      "get": {
        "tags": [
          "report"
        ],
        "summary": "Get year report by sign",
        "description": "This can only be done by the logged in user",
        "operationId": "getReportByTrans",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name":"sign",
            "in":"query",
            "required":true,
            "schema":{
              "type":"string",
              "example":"expenses"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Year report by sign",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "200 OK. Year report success!"
                },
                "results": {
                  "type": "object",
                  "properties": {
                    "month":{
                      "type":"number",
                      "example": 0
                    }
                  }}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string",
              "example": "Not authorized"
            }
          }
        },
        "security": [
          {
            "jwt_token": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "jwt_token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "dsg8v5s8e"
        },
        "name": {
          "type": "string",
          "example": "продукты"
        },
        "sign": {
          "type": "string",
          "enum": [
            "expenses",
            "incomes"
          ],
          "example": "expenses"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Error message"
        }
      }
    }
  }
}